pyproject.toml
[project]
name = "pyglrec"
version = "0.1.0"
description = "An OpenGL frame recorder with NVENC acceleration"
readme = "README.md"
authors = [{ name = "Shinei Arakawa", email = "arakawashinei1115@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "cachetools>=6.2.1",
    "glfw>=2.10.0",
    "imageio[ffmpeg]>=2.37.2",
    "numpy>=2.3.4",
    "pybind11>=3.0.1",
    "pynvvideocodec>=2.0.2",
    "pyopengl>=3.1.10",
    "pyopengl-accelerate>=3.1.10",
]

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
examples = ["click>=8.3.0", "pyglm>=2.8.2"]
